Creating useSearchState to share logic between componendBase and UrlDriven Search
