+++ ui/app/src/components/Search/URLDrivenSearch.tsx
@@ -14,122 +14,60 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import React, { useEffect, useMemo, useRef, useState } from 'react';
 import queryString from 'query-string';
 import { Subject } from 'rxjs';
-import { useActiveSiteId } from '../../utils/hooks/useActiveSiteId';
-import { useEnv } from '../../utils/hooks/useEnv';
-import { useSelection } from '../../utils/hooks/useSelection';
-import { useDispatch } from 'react-redux';
-import { useIntl } from 'react-intl';
 import { useTheme } from '@mui/material/styles';
 import useMediaQuery from '@mui/material/useMediaQuery';
-import { useDetailedItems } from '../../utils/hooks/useDetailedItems';
-import { generateMultipleItemOptions, generateSingleItemOptions, itemActionDispatcher } from '../../utils/itemActions';
-import { search } from '../../services/search';
-import { showErrorDialog } from '../../state/reducers/dialogs/error';
-import { translations } from './translations';
-import { debounceTime, distinctUntilChanged, filter } from 'rxjs/operators';
-import { itemCreated, itemDuplicated, itemsDeleted, itemsPasted, itemUpdated } from '../../state/actions/system';
-import { getHostToHostBus } from '../../modules/Preview/previewContext';
-import { ElasticParams, Filter, MediaItem } from '../../models/Search';
-import { batchActions } from '../../state/actions/misc';
-import { completeDetailedItem } from '../../state/actions/content';
-import { showEditDialog, showItemMegaMenu, showPreviewDialog, updatePreviewDialog } from '../../state/actions/dialogs';
-import { getNumOfMenuOptionsForItem, getSystemTypeFromPath } from '../../utils/content';
-import { AllItemActions, DetailedItem } from '../../models/Item';
-import { getPreviewURLFromPath } from '../../utils/path';
-import { fetchContentXML } from '../../services/content';
-import { actionsToBeShown, initialSearchParameters, setCheckedParameterFromURL, URLDrivenSearchProps } from './utils';
+import { debounceTime, distinctUntilChanged } from 'rxjs/operators';
+import { ElasticParams, Filter } from '../../models/Search';
+import { initialSearchParameters, setCheckedParameterFromURL, URLDrivenSearchProps, useSearchState } from './utils';
 import SearchUI from './SearchUI';
 
 export function URLDrivenSearch(props: URLDrivenSearchProps) {
-  const refs = useRef({ createQueryString: null });
   const { history, location, mode = 'default', onSelect, embedded = false, onAcceptSelection, onClose } = props;
+
+  // region hooks
+  const refs = useRef({ createQueryString: null });
   const queryParams = useMemo(() => queryString.parse(location.search), [location.search]);
   const searchParameters = useMemo(() => setSearchParameters(initialSearchParameters, queryParams), [queryParams]);
-  const [keyword, setKeyword] = useState(queryParams['keywords'] || '');
-  const [currentView, setCurrentView] = useState<'grid' | 'list'>('grid');
-  const [searchResults, setSearchResults] = useState(null);
-  const [selected, setSelected] = useState([]);
   const onSearch$ = useMemo(() => new Subject<string>(), []);
-  const site = useActiveSiteId();
-  const { authoringBase, guestBase } = useEnv();
-  const dispatch = useDispatch();
-  const clipboard = useSelection((state) => state.content.clipboard);
-  const { formatMessage } = useIntl();
-  const [apiState, setApiState] = useState({
-    error: false,
-    errorResponse: null
-  });
-  const [drawerOpen, setDrawerOpen] = useState(window.innerWidth > 960);
-  const [checkedFilters, setCheckedFilters] = useState({});
   const theme = useTheme();
   const desktopScreen = useMediaQuery(theme.breakpoints.up('md'));
-  const [selectedPath, setSelectedPath] = useState(queryParams['path'] as string);
-  const { itemsByPath, isFetching } = useDetailedItems(selected);
+  // endregion
 
-  const selectionOptions = useMemo(() => {
-    if (selected.length === 0) {
-      return null;
-    } else if (selected.length) {
-      if (selected.length === 1) {
-        const path = selected[0];
-        const item = itemsByPath[path];
-        if (item) {
-          return generateSingleItemOptions(item, formatMessage, { includeOnly: actionsToBeShown }).flat();
-        }
-      } else {
-        let items = [];
-        selected.forEach((itemPath) => {
-          const item = itemsByPath[itemPath];
-          if (item) {
-            items.push(item);
-          }
-        });
-        if (items.length && !isFetching) {
-          return generateMultipleItemOptions(items, formatMessage, { includeOnly: actionsToBeShown });
-        }
-      }
-    }
-  }, [formatMessage, isFetching, itemsByPath, selected]);
+  // region state
+  const [keyword, setKeyword] = useState(queryParams['keywords'] || '');
+  const [checkedFilters, setCheckedFilters] = useState({});
+  // endregion
+
+  const {
+    apiState,
+    areAllSelected,
+    selected,
+    currentView,
+    onActionClicked,
+    selectionOptions,
+    onHeaderButtonClick,
+    itemsByPath,
+    handleClearSelected,
+    handleSelect,
+    handleSelectAll,
+    onPreview,
+    guestBase,
+    searchResults,
+    selectedPath,
+    clearPath,
+    onSelectedPathChanges,
+    drawerOpen,
+    toggleDrawer,
+    handleChangeView
+  } = useSearchState({
+    searchParameters,
+    onSelect
+  });
 
   refs.current.createQueryString = createQueryString;
-
-  const refreshSearch = useCallback(() => {
-    search(site, searchParameters).subscribe(
-      (result) => {
-        setSearchResults(result);
-      },
-      (error) => {
-        const { response } = error;
-        if (response && response.response) {
-          setApiState({ error: true, errorResponse: response.response });
-        } else {
-          console.error(error);
-          dispatch(
-            showErrorDialog({
-              error: {
-                message: formatMessage(translations.unknownError)
-              }
-            })
-          );
-        }
-      }
-    );
-  }, [dispatch, formatMessage, searchParameters, site]);
-
-  const handleClearSelected = useCallback(() => {
-    selected.forEach((path) => {
-      onSelect?.(path, false);
-    });
-    setSelected([]);
-  }, [onSelect, selected]);
-
-  useEffect(() => {
-    refreshSearch();
-    return () => setApiState({ error: false, errorResponse: null });
-  }, [refreshSearch]);
 
   useEffect(() => {
     const subscription = onSearch$.pipe(debounceTime(400), distinctUntilChanged()).subscribe((keywords: string) => {
@@ -144,38 +82,12 @@ export function URLDrivenSearch(props: URLDrivenSearchProps) {
   }, [history, onSearch$]);
 
   useEffect(() => {
-    const eventsThatNeedReaction = [
-      itemDuplicated.type,
-      itemsDeleted.type,
-      itemCreated.type,
-      itemUpdated.type,
-      itemsPasted.type
-    ];
-    const hostToHost$ = getHostToHostBus();
-    const subscription = hostToHost$.pipe(filter((e) => eventsThatNeedReaction.includes(e.type))).subscribe(() => {
-      handleClearSelected();
-      refreshSearch();
-    });
-    return () => {
-      subscription.unsubscribe();
-    };
-  }, [handleClearSelected, refreshSearch]);
-
-  useEffect(() => {
     setCheckedFilters(setCheckedParameterFromURL(queryParams));
   }, [queryParams, setCheckedFilters]);
 
   function handleSearchKeyword(keyword: string) {
     setKeyword(keyword);
     onSearch$.next(keyword);
-  }
-
-  function handleChangeView() {
-    if (currentView === 'grid') {
-      setCurrentView('list');
-    } else {
-      setCurrentView('grid');
-    }
   }
 
   function handleFilterChange(filter: Filter, isFilter: boolean) {
@@ -208,11 +120,6 @@ export function URLDrivenSearch(props: URLDrivenSearchProps) {
   function clearFilters() {
     Object.keys(checkedFilters).map((filter) => clearFilter(filter));
     clearPath();
-  }
-
-  function clearPath() {
-    handleFilterChange({ name: 'path', value: undefined }, false);
-    setSelectedPath(undefined);
   }
 
   // createQueryString:
@@ -296,165 +203,6 @@ export function URLDrivenSearch(props: URLDrivenSearchProps) {
     });
   }
 
-  function handleSelect(path: string, isSelected: boolean) {
-    if (isSelected) {
-      // dispatch(completeDetailedItem({ path }));
-      setSelected([...selected, path]);
-    } else {
-      let selectedItems = [...selected];
-      let index = selectedItems.indexOf(path);
-      selectedItems.splice(index, 1);
-      setSelected(selectedItems);
-    }
-    onSelect?.(path, isSelected);
-  }
-
-  function handleSelectAll(checked: boolean) {
-    if (checked) {
-      let selectedItems: any[] = [];
-      searchResults.items.forEach((item: any) => {
-        if (selected.indexOf(item.path) === -1) {
-          selectedItems.push(item.path);
-          onSelect?.(item.path, true);
-        }
-      });
-      setSelected([...selected, ...selectedItems]);
-    } else {
-      let newSelectedItems = [...selected];
-      searchResults.items.forEach((item: any) => {
-        let index = newSelectedItems.indexOf(item.path);
-        if (index >= 0) {
-          newSelectedItems.splice(index, 1);
-          onSelect?.(item.path, false);
-        }
-      });
-      setSelected(newSelectedItems);
-    }
-  }
-
-  function areAllSelected() {
-    if (!searchResults || searchResults.items.length === 0) return false;
-    return !searchResults.items.some((item: any) => !selected.includes(item.path));
-  }
-
-  const onHeaderButtonClick = (event: any, item: MediaItem) => {
-    const path = item.path;
-    dispatch(
-      batchActions([
-        completeDetailedItem({ path }),
-        showItemMegaMenu({
-          path,
-          anchorReference: 'anchorPosition',
-          anchorPosition: { top: event.clientY, left: event.clientX },
-          numOfLoaderItems: getNumOfMenuOptionsForItem({
-            path: item.path,
-            systemType: getSystemTypeFromPath(item.path)
-          } as DetailedItem)
-        })
-      ])
-    );
-  };
-
-  const onPreview = (item: MediaItem) => {
-    const { type, name: title, path: url } = item;
-    switch (type) {
-      case 'Image': {
-        dispatch(
-          showPreviewDialog({
-            type: 'image',
-            title,
-            url
-          })
-        );
-        break;
-      }
-      case 'Page': {
-        dispatch(
-          showPreviewDialog({
-            type: 'page',
-            title,
-            url: `${guestBase}${getPreviewURLFromPath(item.path)}`
-          })
-        );
-        break;
-      }
-      case 'Component':
-      case 'Taxonomy': {
-        dispatch(showEditDialog({ site, path: item.path, authoringBase, readonly: true }));
-        break;
-      }
-      default: {
-        let mode = 'txt';
-        if (type === 'Template') {
-          mode = 'ftl';
-        } else if (type === 'Groovy') {
-          mode = 'groovy';
-        } else if (type === 'JavaScript') {
-          mode = 'javascript';
-        } else if (type === 'CSS') {
-          mode = 'css';
-        }
-        dispatch(
-          showPreviewDialog({
-            type: 'editor',
-            title,
-            url,
-            mode
-          })
-        );
-
-        fetchContentXML(site, url).subscribe((content) => {
-          dispatch(
-            updatePreviewDialog({
-              content
-            })
-          );
-        });
-        break;
-      }
-    }
-  };
-
-  const toggleDrawer = () => {
-    setDrawerOpen(!drawerOpen);
-  };
-
-  const onActionClicked = (option: AllItemActions, event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-    if (selected.length > 1) {
-      const detailedItems = [];
-      selected.forEach((path) => {
-        itemsByPath?.[path] && detailedItems.push(itemsByPath[path]);
-      });
-      itemActionDispatcher({
-        site,
-        item: detailedItems,
-        option,
-        authoringBase,
-        dispatch,
-        formatMessage,
-        clipboard,
-        event
-      });
-    } else {
-      const path = selected[0];
-      const item = itemsByPath?.[path];
-      itemActionDispatcher({
-        site,
-        item,
-        option,
-        authoringBase,
-        dispatch,
-        formatMessage,
-        clipboard,
-        event
-      });
-    }
-  };
-
-  const onSelectedPathChanges = (path: string) => {
-    setSelectedPath(path);
-  };
-
   const onCheckedFiltersChanges = (checkedFilters: object) => {
     setCheckedFilters(checkedFilters);
   };
@@ -481,7 +229,7 @@ export function URLDrivenSearch(props: URLDrivenSearchProps) {
       onSelectedPathChanges={onSelectedPathChanges}
       onCheckedFiltersChanges={onCheckedFiltersChanges}
       apiState={apiState}
-      areAllSelected={areAllSelected()}
+      areAllSelected={areAllSelected}
       guestBase={guestBase}
       handleChangePage={handleChangePage}
       handleChangeRowsPerPage={handleChangeRowsPerPage}
